refactor: StopTimeAdaptor into stoptime package
